[ADD] 반응형 css 추가 및 컴포넌트 분할
- Board에서 BoardHead, BoardTable 나눔
- 검색버튼 추가
- 처음 사이트 접속 시 게시판 이름 안뜨는 문제 수정
- 게시글 없을 시 정보가 없는 상태로 출력하도록 수정
- 모바일 환경에서 검색란 display none
- 테이블 width 수정
- 모바일 환경에서 헤더 box-shadow 삭제
- 닉네임 컴포넌트 수정, 사용자와 비사용자 구별
