+++ src/components/Board/Board.js
@@ -1,74 +1,15 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import { faLock, faEdit } from '@fortawesome/free-solid-svg-icons';
-import {
-  CircularProgress,
-  Table,
-  TableHead,
-  TableBody,
-  TableRow,
-  TableCell,
-  TableContainer,
-  Paper,
-  TextField,
-  Select,
-  MenuItem,
-  Button,
-  Pagination,
-  PaginationItem
-} from '@mui/material';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { CircularProgress, Pagination, PaginationItem } from '@mui/material';
 import qs from 'qs';
-import { postSearchTypeList } from '../../DataExport';
-import Tags from '../Post/Tags';
-import NicknameContainer from '../../containers/Post/NicknameContainer';
 import { MainWrapper } from '../Wrapper/Wrapper';
-import { getTimeForToday } from '../../utils/format';
+import BoardTable from './BoardTable';
+import BoardHead from './BoardHead';
 
 const LoadingCircle = styled(CircularProgress)`
   position: absolute;
   bottom: 50vh;
-`;
-
-const NoneBorderCell = styled(TableCell)`
-  border: none;
-  font-size: 0.75rem;
-  padding: 6px 10px;
-`;
-
-const IconMargin = styled.span`
-  display: inline-block;
-  margin: 2px;
-`;
-
-const InfoIcon = styled(FontAwesomeIcon)`
-  margin: 1px;
-  color: gray;
-`;
-
-const PostTableRow = styled(TableRow)`
-  border-bottom: 1px solid #ddd;
-  background-color: #${props => props.bgcolor};
-`;
-
-const PostNumber = styled.span`
-  font-size: 0.75rem;
-`;
-
-const ReplyCountNumber = styled(PostNumber)`
-  color: gray;
-`;
-
-const PostTitleLink = styled(Link)`
-  display: inline-block;
-  width: 100%;
-  font-size: 0.75rem;
-  text-overflow: ellipsis;
-  vertical-align: middle;
-  color: black;
-  text-decoration: none;
-  cursor: pointer;
 `;
 
 const NoBoardBox = styled.div`
@@ -79,36 +20,6 @@ const NoBoardBox = styled.div`
   margin: 96px 0 96px 0;
 `;
 
-const SearchBar = styled.form`
-  width: 12rem;
-  padding: 4px;
-  margin: 8px;
-  align-items: center;
-`;
-
-const BoardHead = styled.div`
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-`;
-
-const FormSelectField = styled(Select)`
-  margin-right: 2px;
-  width: auto;
-  height: 2rem;
-`;
-
-const BoardHeadRight = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-const BoardHeadLeft = styled.div`
-  font-size: 1.5rem;
-  display: flex;
-  align-items: center;
-`;
-
 const PaginationStyled = styled(Pagination)`
   & ul {
     justify-content: center;
@@ -117,36 +28,6 @@ const PaginationStyled = styled(Pagination)`
       padding: 4px;
     }
   }
-`;
-
-const ButtonStyled = styled(Button)`
-  font-weight: 500;
-  line-height: 1.5;
-  padding: 6px 12px;
-  border-radius: 100px;
-`;
-
-const NumberCol = styled.col`
-  width: 6%;
-  align: center;
-`;
-const TitleCol = styled.col`
-  width: 65%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-`;
-const NicknameCol = styled.col`
-  width: 15%;
-  align: center;
-  text-overflow: ellipsis;
-`;
-const DateCol = styled.col`
-  width: 8%;
-  align: center;
-`;
-const ViewsCountCol = styled.col`
-  width: 6%;
-  align: center;
 `;
 
 const Paginations = props => {
@@ -166,7 +47,7 @@ const Paginations = props => {
     <PaginationStyled
       component="div"
       size="small"
-      count={totalPage}
+      count={totalPage || 1}
       page={boardPage ? parseInt(boardPage, 10) : 1}
       renderItem={item => (
         <PaginationItem component={Link} to={`?${qsMaker(item)}`} {...item} />
@@ -175,175 +56,9 @@ const Paginations = props => {
   );
 };
 
-const PostRow = props => {
-  const { postInfo, boardNameEng, boardPage } = props;
-  const {
-    postId,
-    title,
-    isSecret,
-    nickname,
-    numReply,
-    views,
-    tags,
-    createAt,
-    accountIdString,
-    isNotice
-  } = postInfo;
-
-  const backgroundColor = () => {
-    if (isNotice === 'NOTICE') {
-      return 'eeeeee';
-    }
-    return 'ffffff';
-  };
-
-  const handleTitleLink = () => {
-    if (isSecret === 'NORMAL') {
-      return `/board/${boardNameEng}/${postId}?page=${boardPage}`;
-    }
-    return `/board/${boardNameEng}/${postId}?secret=true&page=${boardPage}`;
-  };
-
-  return (
-    <PostTableRow hover bgcolor={backgroundColor}>
-      <NoneBorderCell nowrap="true" align="center">
-        <PostNumber>{isNotice === 'NOTICE' ? '공지' : postId}</PostNumber>
-      </NoneBorderCell>
-      <NoneBorderCell>
-        <PostTitleLink to={handleTitleLink}>
-          {title}
-          {isSecret === 'SECRET' && (
-            <IconMargin>
-              <InfoIcon icon={faLock} />
-            </IconMargin>
-          )}
-          {numReply ? <ReplyCountNumber> [{numReply}]</ReplyCountNumber> : null}
-          <Tags tags={tags} />
-        </PostTitleLink>
-      </NoneBorderCell>
-      <NoneBorderCell align="center">
-        <NicknameContainer
-          nickname={nickname}
-          accountIdString={accountIdString}
-        />
-      </NoneBorderCell>
-      <NoneBorderCell align="center">
-        {getTimeForToday(createAt)}
-      </NoneBorderCell>
-      <NoneBorderCell align="center">{views}</NoneBorderCell>
-    </PostTableRow>
-  );
-};
-
 const ErrorBoard = props => {
   const { error } = props;
   return <NoBoardBox>{error.message}</NoBoardBox>;
-};
-
-const NoPost = props => {
-  const { keyword } = props;
-  return (
-    <NoBoardBox>{`${keyword}의 검색결과가 하나도 없습니다 😅`}</NoBoardBox>
-  );
-};
-
-const PostListTable = props => {
-  const { res, boardNameEng, boardPage, notice, keyword } = props;
-  const tableColumns = ['번호', '제목', '글쓴이', '작성일', '조회'];
-
-  return (
-    <TableContainer component={Paper}>
-      <Table size="small">
-        <colgroup>
-          <NumberCol />
-          <TitleCol />
-          <NicknameCol />
-          <DateCol />
-          <ViewsCountCol />
-        </colgroup>
-        <TableHead>
-          <TableRow>
-            {tableColumns.map(column => (
-              <NoneBorderCell nowrap="true" align="center" key={column}>
-                {column}
-              </NoneBorderCell>
-            ))}
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {notice.postListItem.content.map(postInfo => (
-            <PostRow
-              key={postInfo.postId}
-              postInfo={postInfo}
-              boardNameEng={boardNameEng}
-              boardPage={boardPage}
-            />
-          ))}
-          {res.postListItem.content.map(postInfo => (
-            <PostRow
-              key={postInfo.postId}
-              postInfo={postInfo}
-              boardNameEng={boardNameEng}
-              boardPage={boardPage}
-            />
-          ))}
-        </TableBody>
-      </Table>
-      {res.postListItem.content.length === 0 ? (
-        <NoPost keyword={keyword} />
-      ) : null}
-    </TableContainer>
-  );
-};
-
-const UpperBar = props => {
-  const {
-    postSearchType,
-    onPostSearchTypeChange,
-    keyword,
-    onSearch,
-    onSearchChange,
-    onWritePost,
-    boardDescription
-  } = props;
-  return (
-    <BoardHead>
-      <BoardHeadLeft>{boardDescription}</BoardHeadLeft>
-      <BoardHeadRight>
-        <FormSelectField
-          margin="dense"
-          value={postSearchType}
-          onChange={onPostSearchTypeChange}
-          variant="standard"
-        >
-          {postSearchTypeList.map(type => (
-            <MenuItem value={type.type} key={type.type}>
-              {type.name}
-            </MenuItem>
-          ))}
-        </FormSelectField>
-        <SearchBar onSubmit={onSearch}>
-          <TextField
-            id="text"
-            type="text"
-            size="small"
-            value={keyword}
-            label="검색"
-            onChange={onSearchChange}
-          />
-        </SearchBar>
-
-        <ButtonStyled
-          variant="contained"
-          size="small"
-          startIcon={<FontAwesomeIcon icon={faEdit} size="sm" />}
-          onClick={onWritePost}
-        >
-          글쓰기
-        </ButtonStyled>
-      </BoardHeadRight>
-    </BoardHead>
-  );
 };
 
 const Board = props => {
@@ -386,7 +101,7 @@ const Board = props => {
   const notice = NoticeData.data;
   return (
     <MainWrapper>
-      <UpperBar
+      <BoardHead
         boardDescription={boardDescription}
         postSearchType={postSearchType}
         onPostSearchTypeChange={onPostSearchTypeChange}
@@ -395,21 +110,18 @@ const Board = props => {
         onSearchChange={onSearchChange}
         onWritePost={onWritePost}
       />
-      <PostListTable
+      <BoardTable
         res={res}
         notice={notice}
         boardNameEng={boardNameEng}
         boardPage={boardPage}
-        keyword={keyword}
       />
-      {res.postListItem.content.length === 0 ? null : (
-        <Paginations
-          res={res}
-          boardNameEng={boardNameEng}
-          boardPage={boardPage}
-          location={location}
-        />
-      )}
+      <Paginations
+        res={res}
+        boardNameEng={boardNameEng}
+        boardPage={boardPage}
+        location={location}
+      />
     </MainWrapper>
   );
 };
